requirements.txt
asgiref
boto3
celery
django-celery-beat
django-celery-results
Django
django_smtp_ssl
django-redis
django-cors-headers
django-coverage-plugin
djangorestframework
django-debug-toolbar
drf-spectacular
drf-spectacular-sidecar
drf_spectacular
django-rest-swagger
redis
pytz
sqlparse
sidecar
typing_extensions
postgres
python-dotenv
python-dateutil
pandas


